pyproject.toml
[tool.poetry]
name = "otel"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
opentelemetry-distro = "^0.49b1"
opentelemetry-exporter-otlp = "^1.28.1"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
